+++ src/XRP/xrp-client.ts
@@ -87,17 +87,17 @@ export default class XrpClient implements XrpClientInterface {
   }
 
   /**
-   * Send the given amount of XRP from the source wallet to the destination Pay ID, allowing
+   * Send the given amount of XRP from the source wallet to the destination PayID, allowing
    * for additional details to be specified for use with supplementary features of the XRP
    * ledger.
    *
-   * @param sendMoneyDetails - a wrapper object containing details for constructing a transaction.
+   * @param sendXrpDetails - a wrapper object containing details for constructing a transaction.
    * @returns A promise which resolves to a string representing the hash of the submitted transaction.
    */
   public async sendWithDetails(
-    sendMoneyDetails: SendXrpDetails,
+    sendXrpDetails: SendXrpDetails,
   ): Promise<string> {
-    return this.decoratedClient.sendWithDetails(sendMoneyDetails)
+    return this.decoratedClient.sendWithDetails(sendXrpDetails)
   }
 
   /**