+++ test/XRP/fakes/fake-xrp-client.ts
@@ -37,7 +37,7 @@ class FakeXrpClient implements XrpClientDecorator {
     return FakeXrpClient.returnOrThrow(this.sendValue)
   }
 
-  async sendWithDetails(_sendMoneyDetails: SendXrpDetails): Promise<string> {
+  async sendWithDetails(_sendXrpDetails: SendXrpDetails): Promise<string> {
     return FakeXrpClient.returnOrThrow(this.sendValue)
   }